// src/components/atoms/AppTypo/interface.ts
export type AppTypoProps = {
  variant:
    | 'h1'
    | 'h2'
    | 'h3'
    | 'p-xl-semi'
    | 'p-xl-med'
    | 'p-xl-reg'
    | 'p-lg-semi'
    | 'p-lg-med'
    | 'p-lg-reg'
    | 'p-md-semi'
    | 'p-md-med'
    | 'p-md-reg'
    | 'p-sm-semi'
    | 'p-sm-med'
    | 'p-sm-reg'
    | 'span'
  children?: React.ReactNode
  style?: object
}